Move benchmark tests to the proper package
